fixed clickthrough behavior
Croblem where you couldn't click through the empty grid cells used to help position the left-hand text. Changed to a single div that mirrors the position of the left-hand column of the 4-column grid by getting the width of the Nav Links div and the position of the main grid on the welcome page.